Add class destructor which calls the chroots nuke() method.
git-svn-id: http://svn.berlios.de/svnroot/repos/fullstory/python-fll/trunk@26117 b7c4aa6e-66fb-0310-853b-84b852d3d4fa

+++ fll/chroot.py
@@ -41,6 +41,9 @@ class FllChroot(object):
         self.path = path
         self.hostname = hostname
 
+    def __del__(self):
+        self.nuke()
+
     def bootstrap(self, bootstrapper='cdebootstrap', suite='sid',
                   flavour='minimal', variant='minbase',
                   arch=None, mirror=None, quiet=False, verbose=False,
@@ -306,7 +309,8 @@ class FllChroot(object):
         self.umountvirtfs()
 
         try:
-            shutil.rmtree(self.path)
+            if os.path.isdir(self.path):
+                shutil.rmtree(self.path)
         except:
             raise FllChrootError('failed to nuke chroot: ' + self.path)
